rewritten request function

+++ src/main/java/OrderBook.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
  * The class is synchronized for concurrency reasons.
  */
 public class OrderBook {
-    private HashMap<Integer, Integer> buyPrices; //from quantity to price
-    private HashMap<Integer, Integer> sellPrices;
+    private HashMap<Double, Double> buyPrices; //from quantity to price
+    private HashMap<Double, Double> sellPrices;
 
     OrderBook(){
         buyPrices = new HashMap<>();
@@ -14,22 +14,22 @@ public class OrderBook {
 
     // Given a midPrice, this function generates an OrderBook of prices and quantities to offer
     public void generateOrderBook(Double midPrice){
-        HashMap<Integer,Integer> newBuyPrices = new HashMap<>();
-        HashMap<Integer,Integer> newSellPrices = new HashMap<>();
+        HashMap<Double,Double> newBuyPrices = new HashMap<>();
+        HashMap<Double,Double> newSellPrices = new HashMap<>();
 
         int i = 0;
-        int startSell = (int) Math.ceil(midPrice);
+        Double startSell = Math.ceil(midPrice);
         while (i < 10){
-            newSellPrices.put(100+i*100, i+startSell);
-            Logger.logPricing("sell", i+startSell, 100+i*100); //logs different prices and quantities being offered
+            newSellPrices.put(100+i*100.0, i+startSell);
+            Logger.logPricing("sell", i+startSell, 100+i*100.0); //logs different prices and quantities being offered
             i++;
         }
 
         int j = 0;
-        int startBuy = (int) Math.floor(midPrice);
+        Double startBuy = Math.floor(midPrice);
         while (j < 10 && startBuy-j > 0){
-            newBuyPrices.put(100+j*100, startBuy-j);
-            Logger.logPricing("buy", startBuy-j,100+j*100);
+            newBuyPrices.put(100+j*100.0, startBuy-j);
+            Logger.logPricing("buy", startBuy-j,100+j*100.0);
             j++;
         }
 
@@ -37,14 +37,13 @@ public class OrderBook {
             buyPrices = newBuyPrices;
             sellPrices = newSellPrices;
         }
-
     }
 
-    public synchronized HashMap<Integer, Integer> getBuyPrices(){
+    public synchronized HashMap<Double, Double> getBuyPrices(){
         return new HashMap(buyPrices); //return copy of hashmap for safety
     }
 
-    public synchronized HashMap<Integer, Integer> getSellPrices(){
+    public synchronized HashMap<Double, Double> getSellPrices(){
         return new HashMap(sellPrices); //same as above
     }
 }
